Refactor works section components
