+++ src/lib/workData.ts
@@ -1,0 +1,172 @@
+export const works = [
+  {
+    title: 'REALIZ-FINANCE 웹',
+    slug: 'realiz-finance',
+    description: '금융시장 리얼타임 데이터 API 스토어',
+    image: '/mock/realiz_finance.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '웹', '기획', '디자인'],
+  },
+  {
+    title: 'REALIZ-FINANCE 백엔드',
+    slug: 'realiz-finance',
+    description: '금융시장 리얼타임 데이터 API 스토어',
+    image: '/realiz.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['백엔드', 'AI', '기획', '디자인'],
+  },
+  {
+    title: 'REALIZ 웹',
+    slug: 'realiz',
+    description: 'AI 모델 마켓플레이스',
+    image: '/mock/realiz.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '웹'],
+  },
+  {
+    title: 'REALIZ 백엔드',
+    slug: 'realiz-server',
+    description: 'AI 모델 마켓플레이스',
+    image: '/realiz.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['백엔드', 'AI'],
+  },
+  {
+    title: '턴업 APP',
+    slug: 'turnup_app',
+    description: '영상, 이미지기반 상품검색 서비스 플랫폼',
+    image: '/mock/turnup_app.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '앱'],
+  },
+  {
+    title: '턴업 브라우저 확장 프로그램',
+    slug: 'turnup_crx',
+    description: '영상, 이미지기반 상품검색 서비스 플랫폼',
+    image: '/mock/turnup_crx.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '브라우저 확장 프로그램'],
+  },
+  {
+    title: '턴업 백오피스 웹',
+    slug: 'turnup_bo',
+    description: '영상, 이미지기반 상품검색 서비스 플랫폼',
+    image: '/realiz.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '웹'],
+  },
+  {
+    title: '턴업 랜딩페이지 v1',
+    slug: 'turn-up',
+    description: '영상, 이미지기반 상품검색 서비스 플랫폼',
+    image: '/mock/turnup_web.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '웹'],
+  },
+  {
+    title: '턴업 서버 & AI 개발',
+    slug: 'turn-up',
+    description: '영상, 이미지기반 상품검색 서비스 플랫폼',
+    image: '/realiz.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['백엔드', 'AI'],
+    filter: ['백엔드', 'AI'],
+  },
+  {
+    title: '라이즈 APP 랜딩페이지',
+    slug: 'rise',
+    description: '라이브커머스 통합 플랫폼',
+    image: '/mock/rise_app_landing.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '앱'],
+  },
+  {
+    title: '라이즈 인사이드 (테스트 프로덕트)',
+    slug: 'rise_inside',
+    description: '턴업 프로젝트의 시장 테스트 프로덕트',
+    image: '/mock/rise_inside.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '앱'],
+  },
+  {
+    title: '라이즈 APP',
+    slug: 'rise',
+    description: '라이브커머스 통합 플랫폼',
+    image: '/mock/rise_app.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '앱'],
+  },
+  {
+    title: '라이즈 백오피스',
+    slug: 'rise',
+    description: '라이브커머스 통합 플랫폼',
+    image: '/realiz.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '웹'],
+  },
+  {
+    title: '라이즈 백엔드',
+    slug: 'rise',
+    description: '라이브커머스 통합 플랫폼',
+    image: '/realiz.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '앱'],
+  },
+  {
+    title: '라이즈 스튜디오 v2',
+    slug: 'rise',
+    description: '라이브커머스 통합 플랫폼',
+    image: '/mock/rise_studio_v2.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '앱'],
+  },
+  {
+    title: '라이즈 스튜디오 v1',
+    slug: 'rise',
+    description: '라이브커머스 통합 플랫폼',
+    image: '/mock/rise_studio_v1.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '앱'],
+  },
+  {
+    title: '패스터',
+    slug: 'faster',
+    description: '동대문 도매상인을 위한 중국 수출입 플랫폼',
+    image: '/mock/faster.png',
+    image2: '/realiz.png',
+    stacks: ['React', 'Next.js', 'TailwindCSS', 'Node.js', 'Express', 'MongoDB'],
+    platforms: ['웹', '모바일', 'CRX', '서버'],
+    filter: ['프론트엔드', '앱'],
+  },
+];
